# docs/api/auth_service_openapi.yaml
openapi: 3.0.3
info:
  title: Authentication Service API
  description: API для автентифікації користувачів та керування токенами
  version: 1.0.0
  
servers:
  - url: http://localhost:8081
    description: Локальний сервер розробки
  - url: http://api.survey-platform.com/auth
    description: Продакшн API сервер
    
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Електронна пошта користувача
        password:
          type: string
          format: password
          description: Пароль, мінімум 8 символів, літери різного регістру, цифри та спеціальні символи
        name:
          type: string
          description: Повне ім'я користувача
      required:
        - email
        - password
        - name

    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: Унікальний ідентифікатор створеного користувача
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - user_id
        - email
        - name

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - email
        - password

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Час дії токену в секундах
        token_type:
          type: string
          enum: [bearer]
        user_id:
          type: string
          format: uuid
      required:
        - access_token
        - refresh_token
        - expires_in
        - token_type
        - user_id

    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token

    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [user, admin, survey_creator]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
        database:
          type: string
          enum: [connected, disconnected]
        details:
          type: string
        service_name:
          type: string
      required:
        - status
        - service_name

paths:
  /auth/register:
    post:
      summary: Реєстрація нового користувача
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Користувач успішно зареєстрований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Неправильний формат даних або користувач з такою електронною поштою вже існує
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Логін користувача
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Користувач успішно автентифікований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Невірні облікові дані
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/refresh:
    post:
      summary: Оновлення токену доступу
      operationId: refreshToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Токен успішно оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Невірний або простроченний токен оновлення
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      summary: Отримання інформації про поточного користувача
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Інформація про користувача
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Невірний або простроченний токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Помилка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      summary: Перевірка стану сервісу
      operationId: checkHealth
      tags:
        - Health
      responses:
        '200':
          description: Сервіс здоровий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Сервіс нездоровий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse' 